+++ app/Http/Controllers/BuildNoteController.php
@@ -4,8 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Build as EloquentBuild;
 use App\Models\Note;
-use App\Services\PageTimer;
-use App\Services\TestingDay;
+use App\Utils\PageTimer;
+use App\Utils\TestingDay;
 use CDash\Model\Build;
 use Illuminate\Http\JsonResponse;
