+++ app/Providers/AuthServiceProvider.php
@@ -7,7 +7,7 @@ namespace App\Providers;
 use App\Models\Test;
 use App\Models\TestImage;
 use App\Models\User;
-use App\Services\ProjectPermissions;
+use App\Utils\ProjectPermissions;
 use CDash\Model\Image;
 use CDash\Model\Project;
 use Illuminate\Auth\Access\Response;
