+++ app/Utils/ProjectPermissions.php
@@ -16,7 +16,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services;
+namespace App\Utils;
 
 use CDash\Model\Project;
 use CDash\Model\UserProject;
