+++ app/cdash/app/Controller/Api/ResultsApi.php
@@ -15,7 +15,7 @@
 =========================================================================*/
 namespace CDash\Controller\Api;
 
-use App\Services\TestingDay;
+use App\Utils\TestingDay;
 
 use CDash\Database;
 use CDash\Model\Build;
