+++ app/cdash/tests/test_autoremovebuilds_on_submit.php
@@ -3,7 +3,7 @@
 // After including cdash_test_case.php, subsequent require_once calls are
 // relative to the top of the CDash source tree
 //
-use App\Services\TestingDay;
+use App\Utils\TestingDay;
 
 use CDash\Database;
 use CDash\Model\BuildGroup;
