+++ app/cdash/xml_handlers/note_handler.php
@@ -16,7 +16,7 @@
 
 require_once 'xml_handlers/abstract_handler.php';
 
-use App\Services\NoteCreator;
+use App\Utils\NoteCreator;
 
 use CDash\Model\Build;
 use App\Models\BuildInformation;
